+++ NEWS.md
@@ -1,7 +1,8 @@
 # osmexport (development version)
 
 * By default, `oexp_download()` now uses a cached download if an export using 
-the same bounding box was downloaded previously.
+the same bounding box was downloaded previously
+* Add tests for `oexp_download()`
 
 # osmexport 0.2.0